Validate the EE definition file earlier
